requirements.txt
setuptools
pytest
pytest-cov
keyboard
nidaqmx
matplotlib
pyserial
pyqt6
pyqtgraph